requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.6
django-allauth==0.58.2
django-cors-headers==4.3.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
flake8==6.0.0
idna==3.7
mccabe==0.7.0
oauthlib==3.2.2
psycopg2==2.9.9
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.12.0
urllib3==2.2.1
